refact(ginkgo-helper): split entry to single file
