+++ go/ginkgo-helper/entry.go
@@ -1,0 +1,22 @@
+package ginkgohelper
+
+import "reflect"
+
+type ContextTableEntryT struct {
+	fmtArgs []any
+	args    []reflect.Value
+}
+
+func (c *ContextTableEntryT) WithFmt(args ...any) *ContextTableEntryT {
+	c.fmtArgs = args
+	return c
+}
+
+func ContextTableEntry(args ...any) *ContextTableEntryT {
+	ret := &ContextTableEntryT{}
+	for i := range args {
+		ret.args = append(ret.args, reflect.ValueOf(args[i]))
+	}
+	ret.fmtArgs = args
+	return ret
+}
