+++ go/ginkgo-helper/table.go
@@ -7,30 +7,23 @@ import (
 	"github.com/onsi/ginkgo/v2"
 )
 
-type ContextTableEntryT struct {
-	fmtArgs []any
-	args    []reflect.Value
-}
-
 func ContextTable(message string, fn any, entries ...*ContextTableEntryT) {
 	for i := range entries {
-		ginkgo.Context(fmt.Sprintf(message, entries[i].fmtArgs...), func() {
-			vfn := reflect.ValueOf(fn)
-			vfn.Call(entries[i].args)
-		})
+		ginkgo.Context(
+			fmt.Sprintf(message, entries[i].fmtArgs...),
+			func() { runContextTableEntry(fn, entries[i]) },
+		)
 	}
 }
 
-func (c *ContextTableEntryT) WithFmt(args ...any) *ContextTableEntryT {
-	c.fmtArgs = args
-	return c
-}
+func runContextTableEntry(fn any, entry *ContextTableEntryT) {
+	vfn := reflect.ValueOf(fn)
+	for j := range entry.args {
+		if entry.args[j].IsValid() {
+			continue
+		}
 
-func ContextTableEntry(args ...any) *ContextTableEntryT {
-	ret := &ContextTableEntryT{}
-	for i := range args {
-		ret.args = append(ret.args, reflect.ValueOf(args[i]))
+		entry.args[j] = reflect.New(vfn.Type().In(j)).Elem()
 	}
-	ret.fmtArgs = args
-	return ret
+	vfn.Call(entry.args)
 }
