+++ go/ginkgo-helper/table_test.go
@@ -14,12 +14,11 @@ func TestHelper(t *testing.T) {
 }
 
 var _ = Describe("ginkgo helper", func() {
-	ContextTable("Given array %v",
-		func(arg []int) {
-			It("array[2]-array[1] should be a equal to 1", func() {
-				Expect(arg[1] - arg[0]).To(Equal(1))
-			})
-		},
+	ContextTable("Given array %v", func(arg []int) {
+		It("array[2]-array[1] should be a equal to 1", func() {
+			Expect(arg[1] - arg[0]).To(Equal(1))
+		})
+	},
 		ContextTableEntry([]int{1, 2}),
 		ContextTableEntry([]int{2, 3}).WithFmt("2 3"),
 	)
